requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.1